requirements.txt
boto==2.49.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.6
django-storages==1.7.2
django-unique-upload==0.2.1
djangorestframework==3.9.1
idna==2.8
Markdown==3.1.1
mysqlclient==1.4.4
Pillow==6.2.1
pkg-resources==0.0.0
python-decouple==3.1
pytz==2019.3
requests==2.22.0
s3utils==0.6.1
sqlparse==0.3.0
urllib3==1.25.7